+ 81c2756edb745bd91ebdc19c09d35966a412189b Do not call getLanguage if not available
+ 71b40910ff2c2531f6f657b4c5b3609710b6e28d Update where join fields in apiv4 call
+ e16d33012d640449aba4653d69c9b9b5f73ce765 Update stats-collector to 1.4.11 to better handle startTimer bug.
+ 4602337982ca37d99c0f88f711987aae97ca8435 Reset unique stats token on each message
+ c8955b7a3beb5f6b1e4fd633c3c66cb6e7394e4a More preferred syntax updates
+ 590eba2418fe8f951e995ef1a018b70e66775016 More deprecated api syntax fixes
+ 783d86457574f5e95542087f2f16926472d1291b New fix for 5.41 fails on render template
+ 68c45778470923be24d75f34ca7d7565294c987a Eliminate need to init() import timer
+ 3f4aa776e5b53e0eff65bd0c382d49e374b9d49a Add unit tests for civicrm import stats patch to confirm stats output.
+ dc4d8ee261c47aeed4cc3b63501d27c28150f074 Update Employer custom field on contact match
+ 9c0a34ff33f7f5e5a95ff787786d324159c21375 Further api fix for 5.41
+ 03181c8041a57023cfeba6cb63c62205c281c555 Remove (int) to stop rounding down of below 1.00 amounts
+ a0208a971a521e64287890869d4e4c2154d6c22e Sync extension with upstream
+ 9626eb19ccffb9e46764ad16953bd9a020d109d9 CiviCRM submodule update

+++ drupal/sites/default/civicrm/extensions/org.wikimedia.smashpig/CRM/Core/Payment/SmashPig.php
@@ -56,7 +56,7 @@ class CRM_Core_Payment_SmashPig extends CRM_Core_Payment {
    */
   public function doPayment(&$params, $component = 'contribute'): array {
     $completedStatusID = CRM_Core_PseudoConstant::getKey('CRM_Contribute_BAO_Contribution', 'contribution_status_id', 'Completed');
-    if ((int) $this->getAmount($params) === 0) {
+    if ((float) $this->getAmount($params) === 0.0) {
       $result['payment_status_id'] = $completedStatusID;
       $result['payment_status'] = 'Completed';
       return $result;
